+++ R/mo_property.R
@@ -145,7 +145,7 @@ mo_fullname <- mo_name
 #' @rdname mo_property
 #' @export
 mo_shortname <- function(x, language = get_locale(), ...) {
-  x.mo <- AMR::as.mo(x, ...)
+  x.mo <- as.mo(x, ...)
   metadata <- get_mo_failures_uncertainties_renamed()
 
   replace_empty <- function(x) {
@@ -223,7 +223,7 @@ mo_type <- function(x, language = get_locale(), ...) {
 #' @rdname mo_property
 #' @export
 mo_gramstain <- function(x, language = get_locale(), ...) {
-  x.mo <- AMR::as.mo(x, ...)
+  x.mo <- as.mo(x, ...)
   metadata <- get_mo_failures_uncertainties_renamed()
 
   x.phylum <- mo_phylum(x.mo)
@@ -290,17 +290,17 @@ mo_rank <- function(x, ...) {
 #' @rdname mo_property
 #' @export
 mo_taxonomy <- function(x, language = get_locale(),  ...) {
-  x <- AMR::as.mo(x, ...)
+  x <- as.mo(x, ...)
   metadata <- get_mo_failures_uncertainties_renamed()
 
-  result <- base::list(kingdom = AMR::mo_kingdom(x, language = language),
-             phylum = AMR::mo_phylum(x, language = language),
-             class = AMR::mo_class(x, language = language),
-             order = AMR::mo_order(x, language = language),
-             family = AMR::mo_family(x, language = language),
-             genus = AMR::mo_genus(x, language = language),
-             species = AMR::mo_species(x, language = language),
-             subspecies = AMR::mo_subspecies(x, language = language))
+  result <- base::list(kingdom = mo_kingdom(x, language = language),
+             phylum = mo_phylum(x, language = language),
+             class = mo_class(x, language = language),
+             order = mo_order(x, language = language),
+             family = mo_family(x, language = language),
+             genus = mo_genus(x, language = language),
+             species = mo_species(x, language = language),
+             subspecies = mo_subspecies(x, language = language))
 
   load_mo_failures_uncertainties_renamed(metadata)
   result
@@ -309,12 +309,12 @@ mo_taxonomy <- function(x, language = get_locale(),  ...) {
 #' @rdname mo_property
 #' @export
 mo_synonyms <- function(x, ...) {
-  x <- AMR::as.mo(x, ...)
+  x <- as.mo(x, ...)
   metadata <- get_mo_failures_uncertainties_renamed()
 
-  IDs <- AMR::mo_property(x = x, property = "col_id", language = NULL)
+  IDs <- mo_property(x = x, property = "col_id", language = NULL)
   syns <- lapply(IDs, function(col_id) {
-    res <- sort(AMR::microorganisms.old[which(AMR::microorganisms.old$col_id_new == col_id), "fullname"])
+    res <- sort(microorganisms.old[which(microorganisms.old$col_id_new == col_id), "fullname"])
     if (length(res) == 0) {
       NULL
     } else {
@@ -335,7 +335,7 @@ mo_synonyms <- function(x, ...) {
 #' @rdname mo_property
 #' @export
 mo_info <- function(x, language = get_locale(),  ...) {
-  x <- AMR::as.mo(x, ...)
+  x <- as.mo(x, ...)
   metadata <- get_mo_failures_uncertainties_renamed()
 
   info <- lapply(x, function(y)
@@ -360,12 +360,12 @@ mo_info <- function(x, language = get_locale(),  ...) {
 #' @importFrom dplyr %>% left_join select mutate case_when
 #' @export
 mo_url <- function(x, open = FALSE, ...) {
-  mo <- AMR::as.mo(x = x, ... = ...)
-  mo_names <- AMR::mo_name(mo)
+  mo <- as.mo(x = x, ... = ...)
+  mo_names <- mo_name(mo)
   metadata <- get_mo_failures_uncertainties_renamed()
 
   df <- data.frame(mo, stringsAsFactors = FALSE) %>%
-    left_join(select(AMR::microorganisms, mo, source, species_id), by = "mo") %>%
+    left_join(select(microorganisms, mo, source, species_id), by = "mo") %>%
     mutate(url = case_when(source == "CoL" ~
                              paste0(gsub("{year}", catalogue_of_life$year, catalogue_of_life$url_CoL, fixed = TRUE), "details/species/id/", species_id),
                            source == "DSMZ" ~
@@ -394,7 +394,7 @@ mo_property <- function(x, property = "fullname", language = get_locale(), ...) 
   if (length(property) != 1L) {
     stop("'property' must be of length 1.")
   }
-  if (!property %in% colnames(AMR::microorganisms)) {
+  if (!property %in% colnames(microorganisms)) {
     stop("invalid property: '", property, "' - use a column name of the `microorganisms` data set")
   }
 
@@ -403,7 +403,7 @@ mo_property <- function(x, property = "fullname", language = get_locale(), ...) 
 
 mo_validate <- function(x, property, ...) {
   
-  load_AMR_package()
+  check_dataset_integrity()
 
   dots <- list(...)
   Becker <- dots$Becker
@@ -417,7 +417,7 @@ mo_validate <- function(x, property, ...) {
 
   # try to catch an error when inputting an invalid parameter
   # so the 'call.' can be set to FALSE
-  tryCatch(x[1L] %in% AMR::microorganisms[1, property],
+  tryCatch(x[1L] %in% microorganisms[1, property],
            error = function(e) stop(e$message, call. = FALSE))
   
   if (is.mo(x) 
@@ -426,7 +426,7 @@ mo_validate <- function(x, property, ...) {
     # this will not reset mo_uncertainties and mo_failures
     # because it's already a valid MO
     x <- exec_as.mo(x, property = property, initial_search = FALSE, ...)
-  } else if (!all(x %in% pull(AMR::microorganisms, property))
+  } else if (!all(x %in% pull(microorganisms, property))
              | Becker %in% c(TRUE, "all")
              | Lancefield %in% c(TRUE, "all")) {
     x <- exec_as.mo(x, property = property, ...)